+++ MANIFEST.in
@@ -1,0 +1,3 @@
+include LICENSE
+include README.md
+include proxyUtil/data/*
