+++ proxyUtil/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+# -*- coding=utf-8 -*-
+
+from .__version__ import __version__
+from .myUtil import *
+from .logFormatter import *
+from .dnsUrl import *
+from .dnsUtil import *
