+++ proxyUtil/__version__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+# -*- coding=utf-8 -*-
+
+__title__ = "proxyUtil"
+__description__ = "some proxy tools"
+__url__ = "https://github.com/mheidari98/proxyUtil"
+__version__ = "0.1.0"
+__author__ = "Mahdi Heidari"
+__author_email__ = "mahdih3idari@gmail.com"
+__license__ = "MIT License"
