+++ requirements.txt
@@ -1,10 +1,30 @@
+anyio==3.6.2
+beautifulsoup4==4.11.1
 certifi==2022.12.7
+cffi==1.15.1
 charset-normalizer==2.1.1
+cryptography==39.0.0
+dnspython==2.3.0
+h11==0.14.0
+h2==4.1.0
+hpack==4.0.0
+httpcore==0.16.3
+httpx==0.23.3
+hyperframe==6.0.1
 idna==3.4
+markdown-it-py==2.1.0
+mdurl==0.1.2
 numpy==1.23.4
 psutil==5.9.2
+pycparser==2.21
+Pygments==2.14.0
 PySocks==1.7.1
 requests==2.28.1
+requests-toolbelt==0.10.1
+rfc3986==1.5.0
+rich==13.3.1
 ruamel.yaml==0.17.21
 ruamel.yaml.clib==0.2.7
+sniffio==1.3.0
+soupsieve==2.3.2.post1
 urllib3==1.26.12
