+++ scripts/sslocal2ssURI
@@ -1,7 +1,10 @@
 #!/usr/bin/env python3
 import argparse
-from modules.myUtil import *
+from proxyUtil import *
 
+ch = logging.StreamHandler()
+ch.setFormatter(CustomFormatter())
+logging.basicConfig(level=logging.ERROR, handlers=[ch])
 
 def main():
     parser = argparse.ArgumentParser(description="ss-local command to shadowsocks URI")
